Issue 6, MultiMap deserializer fix
Moving the testcases to the right test class
